[refactor] add trip plan logic
